+++ map-matcher-library/src/main/java/nu/ndw/nls/routingmapmatcher/util/LineStringScoreUtil.java
@@ -6,46 +6,37 @@ import com.graphhopper.util.PointList;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import nu.ndw.nls.geometry.distance.FractionAndDistanceCalculator;
+import nu.ndw.nls.geometry.distance.FrechetDistanceCalculator;
 import nu.ndw.nls.routingmapmatcher.model.linestring.LineStringLocation;
 import nu.ndw.nls.routingmapmatcher.model.linestring.ReliabilityCalculationType;
 import org.locationtech.jts.geom.CoordinateSequence;
 import org.locationtech.jts.geom.LineString;
 
 @Slf4j
+@RequiredArgsConstructor
 public class LineStringScoreUtil {
 
     private static final boolean REDUCE_TO_SEGMENT = true;
     private static final int MIN_RELIABILITY_SCORE = 0;
     private static final int MAX_RELIABILITY_SCORE = 100;
     private static final double DISTANCE_PENALTY_FACTOR = 1.5;
-    private static final double PATH_LENGTH_DIFFERENCE_PENALTY_FACTOR = 0.1;
 
     private final DistanceCalcCustom distanceCalc = new DistanceCalcCustom();
-    private final FractionAndDistanceCalculator fractionAndDistanceCalculator;
-    private final double absoluteRelativeWeighingFactor;
-
-
-    public LineStringScoreUtil(FractionAndDistanceCalculator fractionAndDistanceCalculator, double absoluteRelativeWeighingFactor) {
-        this.fractionAndDistanceCalculator = fractionAndDistanceCalculator;
-        this.absoluteRelativeWeighingFactor = absoluteRelativeWeighingFactor;
-        log.debug("LineStringScoreUtil created with absoluteRelativeWeighingFactor of {}", absoluteRelativeWeighingFactor);
-    }
+    private final PointListUtil pointListUtil;
+    private final FrechetDistanceCalculator frechetDistanceCalculator;
 
     public double calculateCandidatePathScore(Path path, LineStringLocation lineStringLocation) {
         if (ReliabilityCalculationType.POINT_OBSERVATIONS == lineStringLocation.getReliabilityCalculationType()) {
-            return calculateCandidatePathScoreOnlyPoints(path.calcPoints(), path.getDistance(),
-                    lineStringLocation.getGeometry(), lineStringLocation.getLengthInMeters());
+            return calculateCandidatePathScoreOnlyPoints(path.calcPoints(), lineStringLocation.getGeometry());
         }
-        return calculateCandidatePathScoreLineString(path.calcPoints(), path.getDistance(),
-                lineStringLocation.getGeometry(), lineStringLocation.getLengthInMeters());
+        return calculateCandidatePathScoreLineString(path.calcPoints(), lineStringLocation.getGeometry());
     }
 
     // This method is public to allow applications that implement other mapmatching algorithms than GraphHopper to
     // calculate reliability scores using the same algorithm.
-    public double calculateCandidatePathScoreOnlyPoints(PointList pathPointList, double pathDistance,
-            LineString originalGeometry, Double originalDistance) {
+    public double calculateCandidatePathScoreOnlyPoints(PointList pathPointList, LineString originalGeometry) {
         CoordinateSequence geometryCoordinates = originalGeometry.getCoordinateSequence();
         List<Double> pointDistancesToMatch = new ArrayList<>();
         for (int index = 0; index < geometryCoordinates.size(); index++) {
@@ -54,67 +45,19 @@ public class LineStringScoreUtil {
             pointDistancesToMatch.add(calculateSmallestDistanceToPointList(latitude, longitude, pathPointList));
         }
 
-        double score = MAX_RELIABILITY_SCORE - Collections.min(pointDistancesToMatch)
-                - Collections.max(pointDistancesToMatch);
-
-        if (originalDistance != null) {
-            double pathDistanceLengthDifferenceInMeters = Math.abs(pathDistance - originalDistance);
-            double pathDistanceLengthDifferenceInPercent = pathDistanceLengthDifferenceInMeters / originalDistance;
-            score -= PATH_LENGTH_DIFFERENCE_PENALTY_FACTOR * calculateWeighedScore(pathDistanceLengthDifferenceInMeters,
-                    pathDistanceLengthDifferenceInPercent);
-        }
-
+        double score = MAX_RELIABILITY_SCORE - Collections.min(pointDistancesToMatch) - Collections.max(pointDistancesToMatch);
         return Math.max(MIN_RELIABILITY_SCORE, score);
     }
 
     // This method is public to allow applications that implement other mapmatching algorithms than GraphHopper to
     // calculate reliability scores using the same algorithm.
-    public double calculateCandidatePathScoreLineString(PointList pathPointList, double pathDistance,
-            LineString originalGeometry, Double originalDistance) {
-        double maximumDistanceInMeters = calculateMaximumDistanceInMeters(pathPointList,
-                originalGeometry);
+    public double calculateCandidatePathScoreLineString(PointList pathPointList, LineString originalGeometry) {
+        LineString pathLineString = pointListUtil.toLineString(pathPointList);
+        double maximumDistanceInMeters = frechetDistanceCalculator.calculateFrechetDistanceInMetresFromWgs84(originalGeometry,
+                pathLineString);
 
-        double lengthInMeters = originalDistance != null ? originalDistance
-                : fractionAndDistanceCalculator.calculateLengthInMeters(originalGeometry);
-        double pathDistanceLengthDifferenceInMeters = Math.abs(pathDistance - lengthInMeters);
-        double pathDistanceLengthDifferenceInPercent = pathDistanceLengthDifferenceInMeters / lengthInMeters;
-
-        return Math.max(MIN_RELIABILITY_SCORE, MAX_RELIABILITY_SCORE
-                - (DISTANCE_PENALTY_FACTOR * maximumDistanceInMeters)
-                - (PATH_LENGTH_DIFFERENCE_PENALTY_FACTOR * calculateWeighedScore(pathDistanceLengthDifferenceInMeters,
-                pathDistanceLengthDifferenceInPercent)));
-    }
-
-    private double calculateMaximumDistanceInMeters(PointList pathPointList, LineString geometry) {
-        CoordinateSequence geometryCoordinates = geometry.getCoordinateSequence();
-        double maximumDistanceInMeters = 0.0;
-        for (int index = 0; index < pathPointList.size(); index++) {
-            double latitude = pathPointList.getLat(index);
-            double longitude = pathPointList.getLon(index);
-            double smallestDistanceToLtcLink = calculateSmallestDistanceToCoordinateSequence(latitude, longitude,
-                    geometryCoordinates);
-            maximumDistanceInMeters = Math.max(maximumDistanceInMeters, smallestDistanceToLtcLink);
-        }
-        for (int index = 0; index < geometryCoordinates.size(); index++) {
-            double latitude = geometryCoordinates.getY(index);
-            double longitude = geometryCoordinates.getX(index);
-            double smallestDistanceToLtcLink = calculateSmallestDistanceToPointList(latitude, longitude, pathPointList);
-            maximumDistanceInMeters = Math.max(maximumDistanceInMeters, smallestDistanceToLtcLink);
-        }
-        return maximumDistanceInMeters;
-    }
-
-    private double calculateSmallestDistanceToCoordinateSequence(double latitude, double longitude,
-            CoordinateSequence coordinateSequence) {
-        double smallestDistanceToLtcLink = Double.MAX_VALUE;
-        for (int index = 1; index < coordinateSequence.size(); index++) {
-            double normalizedDistance = distanceCalc.calcNormalizedEdgeDistanceNew(latitude, longitude,
-                    coordinateSequence.getY(index - 1), coordinateSequence.getX(index - 1),
-                    coordinateSequence.getY(index), coordinateSequence.getX(index), REDUCE_TO_SEGMENT);
-            double distanceInMeters = distanceCalc.calcDenormalizedDist(normalizedDistance);
-            smallestDistanceToLtcLink = Math.min(smallestDistanceToLtcLink, distanceInMeters);
-        }
-        return smallestDistanceToLtcLink;
+        double score = MAX_RELIABILITY_SCORE - (DISTANCE_PENALTY_FACTOR * maximumDistanceInMeters);
+        return Math.max(MIN_RELIABILITY_SCORE, score);
     }
 
     private double calculateSmallestDistanceToPointList(double latitude, double longitude, PointList pointList) {
@@ -127,10 +70,5 @@ public class LineStringScoreUtil {
             smallestDistanceToLtcLink = Math.min(smallestDistanceToLtcLink, distanceInMeters);
         }
         return smallestDistanceToLtcLink;
-    }
-
-    private double calculateWeighedScore(double absoluteDifference, double relativeDifference) {
-        return absoluteRelativeWeighingFactor * absoluteDifference + (1.0 - absoluteRelativeWeighingFactor)
-                * relativeDifference;
     }
 }